add needed frameworks
Add MessageKit for fabulous massage animation, JGProgressHUD - has spinner to show when app is thinking and log in/ log out (or sending a big file) could take time, already implement that in log in and registration. Also, add SDWebImage for better cooperation with URL and uploading images. The last one is RealmSwift is build in database as CoreData in Xcode, but have stronger working with cash in app abilities(better for big app)